switch back to images just being thumbnails
